staging: olpc_dcon: drop old VT switch code, watch for CONBLANK event
Eons ago, in a galaxy far far away, Jordan committed code to work around
the fact that X might have put the DCON to sleep and then crashed (in that
galaxy, X crashed a lot; crazy, right?)

This code was based on a custom API.  These days, we have code which watches
for FB unblanks, and should perform the same function.. Therefore, the older
code can be dropped.  We should probably be watching for CONBLANK events to,
so allow those to turn the DCON back on.

Dropping the old code is necessary for building the driver.

Signed-off-by: Andres Salomon <dilinger@queued.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>